LPS-66476 Extract default DDMFormValues creation to a separate class - no logic changes
